+++ packages/moonstone/CHANGELOG.md
@@ -30,6 +30,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/IncrementSlider` so that the knob is spottable via pointer, and 5-way navigation between the knob and the increment/decrement buttons is functional
 - `moonstone/Slider` and `moonstone/IncrementSlider` to not fire `onChange` for value changes from props
+- `moonstone/Popup` and `moonstone/ContextualPopupDecorator` so that spotlight focus cannot escape a popup with a `spotlightRestrict` value of `'self-only'`
 
 ## [1.0.0-beta.1] - 2016-12-30